simplify rpm downloading

+++ contrib/cygwin_build.sh
@@ -1,5 +1,6 @@
 #!/bin/sh
-# Script to cross-compile Julia from Cygwin to MinGW assuming the following packages are installed:
+# Script to cross-compile Julia from Cygwin to MinGW assuming
+# the following packages are installed:
 # - dos2unix
 # - make
 # - wget
@@ -7,23 +8,27 @@
 # - python (for llvm)
 # - mingw64-x86_64-gcc-g++ (for 64 bit)
 # - mingw64-x86_64-gcc-fortran (for 64 bit)
-# - mingw-gcc-g++ (for 32 bit)
-# - mingw-gcc-fortran (for 32 bit)
+# - mingw-gcc-g++ (for 32 bit, not yet tested)
+# - mingw-gcc-fortran (for 32 bit, not yet tested)
 #
-# This script is intended to be executed from the main julia directory as contrib/cygwin_build.sh
+# This script is intended to be executed from the main julia
+# directory as contrib/cygwin_build.sh
 
 # Stop on error
 set -e
 
-# Screen output from all downloads is redirected to a log file to avoid
-# filling up the AppVeyor logs with progress bars
+# Screen output from all downloads is redirected to a log file to
+# avoid filling up the AppVeyor logs with progress bars
 
-# The frequent use of 2>&1 is so AppVeyor doesn't highlight so many normal messages as errors
+# The frequent use of 2>&1 is so AppVeyor doesn't highlight
+# so many normal messages as errors
 if [ -n "`file contrib/relative_path.sh | grep CRLF`" ]; then
-  dos2unix contrib/relative_path.sh deps/jldownload deps/find_python_for_llvm base/version_git.sh 2>&1
+  dos2unix contrib/relative_path.sh deps/jldownload \
+    deps/find_python_for_llvm base/version_git.sh 2>&1
 fi
 
-# Add -C (caching) to CONFIGURE_COMMON in deps/Makefile for slightly faster configure scripts
+# Add -C (caching) to CONFIGURE_COMMON in deps/Makefile for
+# slightly faster configure scripts
 sed -i 's/CONFIGURE_COMMON = --prefix/CONFIGURE_COMMON = -C --prefix/' deps/Makefile
 
 if [ `arch` = x86_64 ]; then
@@ -31,8 +36,11 @@ if [ `arch` = x86_64 ]; then
   echo "override BUILD_MACHINE = x86_64-pc-cygwin" >> Make.user
   
   # Download LLVM binary
-  wget https://sourceforge.net/projects/mingw-w64-dgn/files/others/llvm-3.3-w64-bin-x86_64-20130804.7z > get-deps.log 2>&1
-  bsdtar -xf llvm-3.3-w64-bin-x86_64-20130804.7z
+  f=llvm-3.3-w64-bin-x86_64-20130804.7z
+  if ! [ -e $f ]; then
+    wget https://sourceforge.net/projects/mingw-w64-dgn/files/others/$f > get-deps.log 2>&1
+  fi
+  bsdtar -xf $f
   if [ -d usr ]; then
     for i in bin lib include; do
       mkdir -p usr/$i
@@ -54,7 +62,9 @@ if [ `arch` = x86_64 ]; then
   x86_64-w64-mingw32-ar cr usr/lib/libgtest_main.a
   
   # Download OpenBlas binary
-  wget -O openblas.7z "https://drive.google.com/uc?export=download&id=0B4DmELLTwYmlVWxuTU1QOHozbWM" >> get-deps.log 2>&1
+  if ! [ -e openblas.7z ]; then
+    wget -O openblas.7z "https://drive.google.com/uc?export=download&id=0B4DmELLTwYmlVWxuTU1QOHozbWM" >> get-deps.log 2>&1
+  fi
   bsdtar -xf openblas.7z
   mv lib/libopenblas.dll usr/bin
   chmod +x usr/bin/libopenblas.dll
@@ -64,42 +74,25 @@ if [ `arch` = x86_64 ]; then
   echo "LIBBLASNAME = libopenblas" >> Make.user
   echo 'override LIBLAPACK = $(LIBBLAS)' >> Make.user
   echo 'override LIBLAPACKNAME = $(LIBBLASNAME)' >> Make.user
-  # apparently this openblas library was not build with 64bit integer support
+  # apparently this openblas library was not built with 64bit integer support
   echo "USE_BLAS64 = 0" >> Make.user
   
-  # Download readline binary
-  wget ftp://rpmfind.net/linux/fedora/linux/releases/20/Everything/x86_64/os/Packages/m/mingw64-readline-6.2-3.fc20.noarch.rpm >> get-deps.log 2>&1
-  bsdtar -xf mingw64-readline-6.2-3.fc20.noarch.rpm
+  # Download MinGW binaries from Fedora rpm's for readline,
+  # libtermcap (dependency of readline), pcre, fftw, gmp, mpfr, and zlib
+  for f in readline-6.2-3.fc20 termcap-1.3.1-16.fc20 pcre-8.34-1.fc21 \
+      fftw-3.3.3-2.fc20 gmp-5.1.3-1.fc21 mpfr-3.1.2-1.fc21 zlib-1.2.8-2.fc20; do
+    if ! [ -e mingw64-$f.noarch.rpm ]; then
+      wget ftp://rpmfind.net/linux/fedora/linux/development/rawhide/x86_64/os/Packages/m/mingw64-$f.noarch.rpm >> get-deps.log 2>&1
+    fi
+    bsdtar -xf mingw64-$f.noarch.rpm
+  done
   echo "USE_SYSTEM_READLINE = 1" >> Make.user
   echo "override READLINE = -lreadline -lhistory" >> Make.user
-  # Download libtermcap binary (dependency of readline)
-  wget ftp://rpmfind.net/linux/fedora/linux/releases/20/Everything/x86_64/os/Packages/m/mingw64-termcap-1.3.1-16.fc20.noarch.rpm >> get-deps.log 2>&1
-  bsdtar -xf mingw64-termcap-1.3.1-16.fc20.noarch.rpm
-  
-  # Download pcre binary
-  wget ftp://rpmfind.net/linux/fedora/linux/releases/18/Everything/x86_64/os/Packages/m/mingw64-pcre-8.31-1.fc18.noarch.rpm >> get-deps.log 2>&1
-  bsdtar -xf mingw64-pcre-8.31-1.fc18.noarch.rpm
   echo "USE_SYSTEM_PCRE = 1" >> Make.user
   echo "override PCRE_CONFIG = $PWD/usr/bin/pcre-config" >> Make.user
-  
-  # Download fftw binary
-  wget ftp://rpmfind.net/linux/fedora/linux/releases/20/Everything/x86_64/os/Packages/m/mingw64-fftw-3.3.3-2.fc20.noarch.rpm >> get-deps.log 2>&1
-  bsdtar -xf mingw64-fftw-3.3.3-2.fc20.noarch.rpm
   echo "USE_SYSTEM_FFTW = 1" >> Make.user
-  
-  # Download gmp binary
-  wget ftp://rpmfind.net/linux/fedora/linux/development/rawhide/x86_64/os/Packages/m/mingw64-gmp-5.1.3-1.fc21.noarch.rpm >> get-deps.log 2>&1
-  bsdtar -xf mingw64-gmp-5.1.3-1.fc21.noarch.rpm
   echo "USE_SYSTEM_GMP = 1" >> Make.user
-  
-  # Download mpfr binary
-  wget ftp://rpmfind.net/linux/fedora/linux/development/rawhide/x86_64/os/Packages/m/mingw64-mpfr-3.1.2-1.fc21.noarch.rpm >> get-deps.log 2>&1
-  bsdtar -xf mingw64-mpfr-3.1.2-1.fc21.noarch.rpm
   echo "USE_SYSTEM_MPFR = 1" >> Make.user
-  
-  # Download zlib binary
-  wget ftp://rpmfind.net/linux/fedora/linux/releases/20/Everything/x86_64/os/Packages/m/mingw64-zlib-1.2.8-2.fc20.noarch.rpm >> get-deps.log 2>&1
-  bsdtar -xf mingw64-zlib-1.2.8-2.fc20.noarch.rpm
   echo "USE_SYSTEM_ZLIB = 1" >> Make.user
   
   # Move all downloaded bin, lib, and include files into build tree
@@ -108,6 +101,9 @@ if [ `arch` = x86_64 ]; then
   if ! [ -d usr/include/readline ]; then
     mv usr/x86_64-w64-mingw32/sys-root/mingw/include/* usr/include
   fi
+
+  # Modify prefix in pcre-config
+  sed -i "s|prefix=/usr/x86_64-w64-mingw32/sys-root/mingw|prefix=$PWD/usr|" usr/bin/pcre-config
   
   # Rename versioned dll's
   mv usr/bin/libpcre-1.dll usr/bin/libpcre.dll
@@ -126,7 +122,6 @@ if [ -z `which gcc 2>/dev/null` ]; then
   echo 'override HOSTCC = $(CROSS_COMPILE)gcc' >> Make.user
 fi
 
-#make -C deps getall >> get-deps.log 2>&1
 make -C deps get-uv get-double-conversion get-openlibm get-openspecfun get-random \
   get-suitesparse get-arpack get-unwind get-osxunwind get-patchelf get-utf8proc >> get-deps.log 2>&1
